Introduce ValidationFunc and move existing validation code into one
